feat: refactor styles and update imports for screens
