+++ src/screens/ErrorScreen.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, Text, TouchableOpacity } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import AppHeader from '../components/AppHeader/AppHeader';
-import { palette } from '../Styles/Palette';
+import { errorScreenStyles } from '../styles/screens/ErrorScreen.styles';
 import type { SavedLocation } from '../store/useLocationStore';
 import type { LocationWeatherState } from '../hooks/useMultiLocationWeather';
 
@@ -26,8 +26,8 @@ export default function ErrorScreen({
   locationLoadingStates,
 }: ErrorScreenProps) {
   return (
-    <SafeAreaView style={styles.container}>
-      <View style={styles.errorContainer}>
+    <SafeAreaView style={errorScreenStyles.container}>
+      <View style={errorScreenStyles.errorContainer}>
         <AppHeader
           onSettingsPress={onSettingsPress}
           savedLocations={savedLocations}
@@ -35,15 +35,15 @@ export default function ErrorScreen({
           onLocationSelect={onLocationSelect}
           locationLoadingStates={locationLoadingStates}
         />
-        <View style={styles.errorContent}>
-          <Text style={styles.errorTitle}>Unable to Load Weather</Text>
-          <Text style={styles.errorMessage}>{errorMessage}</Text>
-          <TouchableOpacity style={styles.retryButton} onPress={onRetry}>
-            <Text style={styles.retryButtonText}>Retry</Text>
+        <View style={errorScreenStyles.errorContent}>
+          <Text style={errorScreenStyles.errorTitle}>Unable to Load Weather</Text>
+          <Text style={errorScreenStyles.errorMessage}>{errorMessage}</Text>
+          <TouchableOpacity style={errorScreenStyles.retryButton} onPress={onRetry}>
+            <Text style={errorScreenStyles.retryButtonText}>Retry</Text>
           </TouchableOpacity>
-          <Text style={styles.supportText}>
+          <Text style={errorScreenStyles.supportText}>
             If this happens often, please contact support at{'\n'}
-            <Text style={styles.supportEmail}>{process.env.EXPO_PUBLIC_SUPPORT_EMAIL || 'support@eladkarni.solutions'}</Text>
+            <Text style={errorScreenStyles.supportEmail}>{process.env.EXPO_PUBLIC_SUPPORT_EMAIL || 'support@eladkarni.solutions'}</Text>
           </Text>
         </View>
       </View>
@@ -51,54 +51,4 @@ export default function ErrorScreen({
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: palette.primaryDark,
-  },
-  errorContainer: {
-    flex: 1,
-  },
-  errorContent: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-  },
-  errorTitle: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    color: palette.textColor,
-    marginBottom: 12,
-    textAlign: 'center',
-  },
-  errorMessage: {
-    fontSize: 16,
-    color: palette.highlightColor,
-    textAlign: 'center',
-    marginBottom: 24,
-    paddingHorizontal: 20,
-  },
-  retryButton: {
-    backgroundColor: palette.primaryColor,
-    paddingVertical: 12,
-    paddingHorizontal: 32,
-    borderRadius: 8,
-  },
-  retryButtonText: {
-    color: palette.textColor,
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  supportText: {
-    fontSize: 12,
-    color: palette.highlightColor,
-    textAlign: 'center',
-    marginTop: 32,
-    opacity: 0.7,
-  },
-  supportEmail: {
-    color: palette.primaryColor,
-    fontWeight: '500',
-  },
-});
+
