+++ src/styles/screens/SkeletonScreen.styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+import { palette } from '../../Styles/Palette';
+
+export const skeletonScreenStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: palette.primaryDark,
+  },
+});
