+++ src/styles/screens/WeatherScreen.styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+import { palette } from '../../Styles/Palette';
+
+export const weatherScreenStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 5,
+    backgroundColor: palette.primaryDark,
+  },
+});
